♻️ Sidebar 내에서 사용할 Mogaco 컴포넌트 리팩토링
- 폴더 구조 정리 (Sidebar의 children으로 사용할 아이템은 Sidebar/Contents에서 관리하도록)
- Sidebar/Contents/Mogaco 컴포넌트의 TitleWrapper/InfoWrapper 분리
