+++ app/frontend/src/components/Sidebar/Contents/Mogaco/index.css.ts
